2 Point charges
If we have two point charges, we can fix one of them and look for the
best point for second one ann repeat this procedure for second point.
in this code, we start from 24 starting points and for each we do this
process 6 times.
Run Main and Attached functions:
pointchargefn1+pointchargefn2+evalpotentioal